set overall assesment float

+++ db/migrate/20160804095116_add_overallassesment_to_interviewerfeedback.rb
@@ -1,5 +1,5 @@
 class AddOverallassesmentToInterviewerfeedback < ActiveRecord::Migration
   def change
-    add_column :interviewerfeedbacks, :overallassesment, :integer
+    add_column :interviewerfeedbacks, :overallassesment, :double
   end
 end
